Make the facades a first class concept.
- Eleveta the facade types to be visible in the public API
surface and rename them from Facade* to Fz*. The purpose here
is to enable better ergonomics for all the cases where the keys
are not known at compile time, such as for deserialization.

- Implement the From and FromIterator traits on the facede types
for better usability.

- Changed the new constructor for the facade sets to take a vector
of items instead of a facade maps. The facade maps can now be used
with the From trait instead.

- Added the `macros` cargo feature to control the inclusion of the
procedural macros.