Major update (#51)
- Improved usability by implementing all the methods that
were previously just on the traits as normal methods on the
collections themselves. This avoids the need to import
the traits to use the collections, making them more user-friendly.

- Revamped the FzStringMap/Set types. Their implementation
is now simpler, yet the API is more flexible.

- Tidied up a lot of generic bounds. Many of the
bounds are removed, making the types easier to use

- SetQuery and MapQuery now have one fewer generics,
making them considerably easier and natural to use.

- Enabled more lints and fixed the resulting warnings.